Fix Content-Length format specifier for size_t

+++ chap07/web_server.c
@@ -247,7 +247,7 @@ void serve_resource(struct client_info *client, const char *path) {
     sprintf(buffer, "Connection: close\r\n");
     send(client->socket, buffer, strlen(buffer), 0);
 
-    sprintf(buffer, "Content-Length: %u\r\n", cl);
+    sprintf(buffer, "Content-Length: %lu\r\n", (unsigned long)cl);
     send(client->socket, buffer, strlen(buffer), 0);
 
     sprintf(buffer, "Content-Type: %s\r\n", ct);
